refactor(painters/dom): extract drawing fragment renderer
Move renderDrawingFragment and isHeaderWordArtWatermark from
DomPainter into a standalone drawings/renderDrawingFragment.ts
module, drop the redundant renderDrawingContent wrapper method,
and pass the painter's frame/anchor helpers in via parameters.
